--- src/components/carbon/LeftNav/index.js
@@ -1,3 +1,0 @@
-import LeftNav from './LeftNav';
-
-export default LeftNav;
